refactor: Preact CLI tests
CR changes
