+++ tests/async-test.js
@@ -1,19 +1,23 @@
 import test from 'tape';
 
-const testTimeout = 30 * 1000;
+const defaultTestTimeout = 10 * 1000;
 
-const asyncTest = (name, fn) => {
-	test(name, (t) => {
-		t.timeoutAfter(testTimeout);
+const asyncTest = (...args) => {
+	let { fn, tapeArgs, hasTimeout } = parseArgs(args);
+
+	test(...tapeArgs, (t) => {
+		if (!hasTimeout) t.timeoutAfter(defaultTestTimeout);
 		fn(t)
 			.then(() => t.end())
 			.catch(e => t.end(e));
 	});
 };
 
-asyncTest.only = (name, fn) => {
-	test.only(name, t => {
-		t.timeoutAfter(testTimeout);
+asyncTest.only = (...args) => {
+	let { fn, tapeArgs, hasTimeout } = parseArgs(args);
+
+	test.only(...tapeArgs, t => {
+		if (!hasTimeout) t.timeoutAfter(defaultTestTimeout);
 		fn(t)
 			.then(() => t.end())
 			.catch(e => t.end(e));
@@ -21,3 +25,15 @@ asyncTest.only = (name, fn) => {
 };
 
 export default asyncTest;
+
+const parseArgs = args => {
+	let fn = args[args.length - 1];
+	let tapeArgs = args.slice(0, args.length - 1);
+	let opts = tapeArgs[1];
+
+	return {
+		fn,
+		tapeArgs,
+		hasTimeout: opts && opts.timeout !== undefined
+	};
+};
