+++ tests/lib/output.js
@@ -7,10 +7,5 @@ const rm = promisify(rimraf);
 
 export const outputPath = resolve(__dirname, '../output');
 
-export const setup = async () => {
-	await mkdirp(outputPath);
-};
-
-export const clean = async () => {
-	await rm(outputPath);
-};
+export const setup = () => mkdirp(outputPath);
+export const clean = () => rm(outputPath);
